+++ src/config/yaml/mod.rs
@@ -9,44 +9,61 @@ use std::fs::File;
 use std::io::ErrorKind;
 use std::path::{Path, PathBuf};
 
-pub struct Projects(pub HashMap<PathBuf, Project>);
+pub struct Config {
+    pub root_project_dir: PathBuf,
+    pub projects: HashMap<PathBuf, Project>,
+}
 
-impl Projects {
+impl Config {
     pub fn load(root_project_dir: &Path) -> Result<Self> {
+        let root_project_dir = canonicalize_dir(root_project_dir)?;
+
         let mut projects = HashMap::new();
 
-        fn add_project(project_dir: &Path, projects: &mut HashMap<PathBuf, Project>) -> Result<()> {
-            let project_dir = project_dir.canonicalize().map_err(|e| {
-                let context = if e.kind() == ErrorKind::NotFound {
-                    format!("Directory {} does not exist", project_dir.display())
-                } else {
-                    format!("Invalid directory: {}", project_dir.display())
-                };
-                Error::new(e).context(context)
-            })?;
-
+        fn add_project(
+            project_dir: PathBuf,
+            projects: &mut HashMap<PathBuf, Project>,
+        ) -> Result<()> {
             if projects.contains_key(&project_dir) {
                 return Ok(());
             }
 
-            let project = Projects::load_project(&project_dir)?;
-            let import_paths: Vec<_> = project
+            let project = Config::load_project(&project_dir)?;
+
+            let import_paths = project
                 .imports
                 .iter()
-                .map(|import| project_dir.join(import))
-                .collect();
+                .map(|(import_name, import_dir)| {
+                    canonicalize_dir(&project_dir.join(import_dir))
+                        .map(|dir| (import_name.clone(), dir))
+                })
+                .collect::<Result<Vec<_>>>()?;
             projects.insert(project_dir, project);
 
-            for import_path in import_paths {
-                add_project(&import_path, projects)?;
+            for (import_name, import_dir) in import_paths {
+                add_project(import_dir.clone(), projects)
+                    .and_then(|_| match &projects[&import_dir].name {
+                        None => Err(anyhow::anyhow!(
+                            "Project cannot be imported as it has no name"
+                        )),
+                        Some(name) if name != &import_name => Err(anyhow::anyhow!(
+                            "The project should be imported with name {}",
+                            name
+                        )),
+                        _ => Ok(()),
+                    })
+                    .with_context(|| format!("Failed to import {}", &import_name))?;
             }
 
             Ok(())
         }
 
-        add_project(root_project_dir, &mut projects)?;
+        add_project(root_project_dir.clone(), &mut projects)?;
 
-        Ok(Self(projects))
+        Ok(Self {
+            root_project_dir,
+            projects,
+        })
     }
 
     fn load_project(project_dir: &Path) -> Result<Project> {
@@ -56,6 +73,15 @@ impl Projects {
         })?;
         let project: Project = serde_yaml::from_reader(config_file)
             .with_context(|| format!("Invalid format for {}", config_file_path.display()))?;
+
+        if let Some(project_name) = &project.name {
+            if !is_valid_project_name(&project_name) {
+                return Err(anyhow::anyhow!(
+                    "{} is not a valid project name",
+                    project_name
+                ));
+            }
+        }
 
         if let Some(invalid_target_name) = project
             .targets
@@ -72,8 +98,19 @@ impl Projects {
     }
 
     pub fn get_project_dirs(&self) -> Vec<PathBuf> {
-        self.0.keys().cloned().collect()
+        self.projects.keys().cloned().collect()
     }
+}
+
+fn canonicalize_dir(dir: &Path) -> Result<PathBuf> {
+    dir.canonicalize().map_err(|e| {
+        let context = if e.kind() == ErrorKind::NotFound {
+            format!("Directory {} does not exist", dir.display())
+        } else {
+            format!("Invalid directory: {}", dir.display())
+        };
+        Error::new(e).context(context)
+    })
 }
 
 pub fn is_valid_target_name(target_name: &str) -> bool {
@@ -81,6 +118,13 @@ pub fn is_valid_target_name(target_name: &str) -> bool {
         static ref RE: Regex = Regex::new(r"^\w[-\w]*$").unwrap();
     }
     RE.is_match(target_name)
+}
+
+pub fn is_valid_project_name(project_name: &str) -> bool {
+    lazy_static! {
+        static ref RE: Regex = Regex::new(r"^\w[-\w]*$").unwrap();
+    }
+    RE.is_match(project_name)
 }
 
 #[cfg(test)]