+++ src/domain.rs
@@ -1,3 +1,4 @@
+use std::fmt;
 use std::path::PathBuf;
 
 pub type TargetId = usize;
@@ -6,10 +7,25 @@ pub type TargetId = usize;
 pub struct Target {
     pub id: TargetId,
     pub name: String,
+    pub project: Project,
     pub dependencies: Vec<TargetId>,
-    pub path: PathBuf,
     pub input_paths: Vec<PathBuf>,
     pub output_paths: Vec<PathBuf>,
     pub build: Option<String>,
     pub service: Option<String>,
+}
+
+impl fmt::Display for Target {
+    fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
+        if let Some(project_name) = &self.project.name {
+            fmt.write_fmt(format_args!("{}::", project_name))?;
+        }
+        fmt.write_str(&self.name)
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct Project {
+    pub dir: PathBuf,
+    pub name: Option<String>,
 }